+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2021 - 2021/03/01
+- Tidied up
+- Added `--zero/-0` flag to return non-zero exit code when an incompatible
+  license is found
+
 ## 2021 - 2021/01/24
 - Updated requirements
 - Fallback to requirements.txt when poetry throws an error and direct the user